Move viewport to whiteboard
