+++ client/src/whiteboard/drawing/rectangle.ts
@@ -1,5 +1,5 @@
 import { server } from "../../server";
-import { viewport } from "../../viewport";
+import { viewport } from "../viewport";
 import drawUtil from "./util";
 
 const layer = document.getElementById("whiteboard-drawing-layer") as unknown as SVGSVGElement;
