+++ client/src/whiteboard/drawing/util.ts
@@ -1,5 +1,5 @@
 import { grid } from "../../grid";
-import { viewport } from "../../viewport";
+import { viewport } from "../viewport";
 
 const cursor = document.getElementById("whiteboard-drawing-cursor") as unknown as SVGCircleElement;
