+++ client/src/whiteboard/viewport.ts
